refactor: re-use genericFetchAllPages function across different views
